backend/requirements.txt
# SmartFix-AI Backend Dependencies

# Core Framework
fastapi==0.103.1
uvicorn==0.23.2
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic==2.4.2
pydantic-settings==2.10.1

# Database & Caching
sqlalchemy==2.0.23
redis==5.0.1

# HTTP Client
requests==2.31.0
aiohttp==3.9.1

# AI & ML Services
google-generativeai==0.3.1
transformers==4.36.2
torch>=2.2.0
sentence-transformers==2.2.2
huggingface-hub==0.19.4

# Image Processing & OCR
pillow==10.0.1
pytesseract==0.3.10
opencv-python==4.8.0.74

# Text Processing & NLP
nltk==3.8.1
spacy==3.7.2
scikit-learn==1.3.2
textblob==0.17.1

# External Services
twilio==8.5.0

# System Monitoring (Windows)
wmi==1.5.1; sys_platform == "win32"

# Utilities
numpy==1.24.3
psutil==5.9.6

# Development & Testing (Optional)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2